\name{actionButton2}
\alias{actionButton2}
\title{actionButton2}
\usage{
  actionButton2(inputId, label,
    class = "btn action-button")
}
\arguments{
  \item{inputId}{Specifies the input slot that will be used
  to access the value.}

  \item{label}{The contents of the button–usually a text
  label, but you could also use any other HTML, like an
  image.}

  \item{class}{the class parameters of the
  \code{actionButton}}
}
\description{
  an extension of actionButton
}

