\name{mini_plot}
\alias{mini_plot}
\title{Mini plot}
\usage{
  mini_plot(title, subtitle = "", vec, sel, vec2 = NULL)
}
\arguments{
  \item{title}{plot title}

  \item{vec}{data vector, e.g. show density given all data}

  \item{sel}{user input; vector of length two (c(min,
  max))}

  \item{vec2}{secondary data vector, e.g. show density
  given specific selection}
}
\description{
  ...
}

